feat: enforce 5-10 minute breaks at the end of sessions
Sessions that are up to and including 60 minutes are shortened by 5 minutes and
those that are longer than that get shortened by 10 minutes.

For simplicity this is implemented by displaying the shortened session times
everywhere where the times are visible to the user. In the database everything
is still stored with the full duration (including break). In particular, this
avoids any issues with overlapping sessions (even though that should not happen
anyway because start time should always be the full hour or half past).

closes #82